+++ spec/unit/resources/matchers_spec.rb
@@ -51,8 +51,10 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_fcoe_network_if_missing('')
     expect(chef_run).to_not delete_oneview_fcoe_network('')
 
-    # oneview_firmware_bundle
-    expect(chef_run).to_not add_oneview_firmware_bundle('')
+    # oneview_firmware
+    expect(chef_run).to_not add_oneview_firmware('')
+    expect(chef_run).to_not remove_oneview_firmware('')
+    expect(chef_run).to_not create_oneview_firmware_custom_spp('')
 
     # oneview_interconnect
     expect(chef_run).to_not set_oneview_interconnect_uid_light('')
@@ -76,6 +78,12 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_logical_switch_group_if_missing('')
     expect(chef_run).to_not delete_oneview_logical_switch_group('')
 
+    # oneview_logical_switch
+    expect(chef_run).to_not create_oneview_logical_switch('')
+    expect(chef_run).to_not create_oneview_logical_switch_if_missing('')
+    expect(chef_run).to_not delete_oneview_logical_switch('')
+    expect(chef_run).to_not refresh_oneview_logical_switch('')
+
     # oneview_network_set
     expect(chef_run).to_not create_oneview_network_set('')
     expect(chef_run).to_not create_oneview_network_set_if_missing('')
@@ -88,7 +96,6 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not remove_oneview_rack_from_rack('')
 
     # oneview_server_hardware
-    expect(chef_run).to_not add_oneview_server_hardware('')
     expect(chef_run).to_not add_oneview_server_hardware_if_missing('')
     expect(chef_run).to_not remove_oneview_server_hardware('')
     expect(chef_run).to_not refresh_oneview_server_hardware('')