+++ CHANGELOG.md
@@ -13,8 +13,11 @@
   - Added support to Server hardware actions
   - Added support to Interconnect actions
   - Added support to Logical switch group actions
+  - Added support to Logical switch actions
   - Added support to Firmware bundle actions
   - Added support to Switch actions
+  - Added support to Firmware driver actions and integrated it with firmware bundles in a resource called Firmware
+  - Added diff output for updating resources
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0